TaskManager/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-crispy-forms==1.14.0
filepath==0.1
greenlet==2.0.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.23.5
pandas==1.5.2
Pillow==9.3.0
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
tzdata==2022.6
